user is None jwt login redirection

+++ pong/friend/views.py
@@ -4,16 +4,24 @@ from django.contrib.auth.models import User
 from django.contrib import messages
 from django.db.models import Q
 from django.utils import timezone
+from django.urls import reverse
 from .models import Friendship, Player
+from player.jwt import generate_jwt, decode_jwt, token_user, set_jwt_token
 import logging
 
 
 @login_required
 def index(request):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     return render(request, "friend/index.html")
 
 @login_required
 def add(request):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     if request.method == 'POST':
         you = request.user
         you.last_login = timezone.now()
@@ -38,12 +46,18 @@ def add(request):
 
 @login_required    
 def delete_friend(request, id_friendship):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     request = Friendship.objects.get(id=id_friendship)
     request.delete()
     return redirect("friend:list")
 
 @login_required
 def accept(request, id_friendship):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     request = Friendship.objects.get(id=id_friendship)
     request.status = 'accepted'
     request.save()
@@ -51,6 +65,9 @@ def accept(request, id_friendship):
 
 @login_required
 def refuse(request, id_friendship):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     request = Friendship.objects.get(id=id_friendship)
     request.status = 'refused'
     request.save()
@@ -58,6 +75,9 @@ def refuse(request, id_friendship):
 
 @login_required
 def list(request):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     you = request.user
     friends = Friendship.objects.filter(Q(user=you, status='accepted') | Q(friend=you, status='accepted'))
     you.last_login = timezone.now()
@@ -66,6 +86,9 @@ def list(request):
 
 @login_required
 def pending(request):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     you = request.user
     friends = Friendship.objects.filter(Q(friend=you, status='pending'))
     you.last_login = timezone.now()
@@ -74,6 +97,9 @@ def pending(request):
 
 @login_required
 def refused(request):
+    user = token_user(request)
+    if user is None: 
+        return redirect(reverse('player:login'))
     you = request.user
     friends = Friendship.objects.filter(Q(friend=you, status='refused'))
     you.last_login = timezone.now()
